moved config python file into handlers folder to avoid confusion with other config dir
