--- NRWAL/config/README.rst
@@ -1,10 +1,0 @@
-****************************
-NRWAL Configuration Software
-****************************
-
-This directory contains the python software to handle the NRWAL configuration
-yaml files. Were you looking for the actual NRWAL cost model configurations?
-If so, check out these other directories in the library:
-
- - `The NRWAL Equation Library <https://github.com/NREL/NRWAL/tree/main/NRWAL/analysis_library>`_.
- - `Default NRWAL Configs <https://github.com/NREL/NRWAL/tree/main/NRWAL/default_configs>`_.
