--- NRWAL/config/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-NRWAL config framework.
-"""
-from .config import NrwalConfig
