+++ NRWAL/handlers/__init__.py
@@ -2,6 +2,7 @@
 """
 Handler objects to interface with NRWAL equation library.
 """
+from NRWAL.handlers.config import NrwalConfig
 from NRWAL.handlers.equations import Equation
 from NRWAL.handlers.groups import EquationGroup, VariableGroup
 from NRWAL.handlers.directories import EquationDirectory
